nemo_mqtt/mqtt_requirements.txt
# Additional requirements for MQTT plugin with Redis integration
redis>=4.0.0
paho-mqtt>=1.6.0
